[Modify] DevloperIntroduceController
