+++ SMS_iOS_V1/SMS/SMS/View/CollectionViewCell/MyPage/DevCollectionViewCell.swift
@@ -4,4 +4,11 @@ class DevCollectionViewCell: UICollectionViewCell {
     @IBOutlet weak var nameLbl: UILabel!
     @IBOutlet weak var imageView: UIImageView!
     @IBOutlet weak var partLbl: UILabel!
+    
+    func setting(_ people: People) {
+        imageView.image = UIImage(named: people.image)
+        imageView.layer.cornerRadius = imageView.frame.height / 1.95
+        nameLbl.text = people.name
+        partLbl.text = people.part
+    }
 }
